Fix issue #2 - tags with spaces break the lists.
THe code had a nested range to try to find the pages based fact that the
title of the list page was the tag needed. But you can't depend on the
capitalization be correct. Hugo has already done the hard work for us,
so use .Data.Pages instead.

While at it, consolidate the tag/archive layouts since with this change,
they are the same, anyway.