support different names for social icons
THe code was making the assumption that the name of the icon for a
social media service was exaclty the same as the name for the service.
That isn't always true. And even when it is, you might want to use an
alternate icon. Create a data file that maps the name of the service
into the icon font class name. The current implementation assumes you
are using fontawesome icon fonts.